pyproject.toml
[project]
name = "chat-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.111.0",
    "google-generativeai>=0.5.0",
    "langchain>=0.2.0,<0.3.0",
    "langchain-community>=0.2.0,<0.3.0",
    "langchain-core>=0.2.0,<0.3.0",
    "langchain-groq>=0.1.0,<0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-pinecone>=0.1.0",
    "pinecone-client>=4.1.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.26",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.56.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.30.0",
]
